+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 Changelog
 =========
 
+<<<<<<< HEAD
 # 1.1.1
 
 - add support for `bpmn:dataStore`
 - remove dataStore prop if none
+=======
+# 2.0.0
+
+Handle participants add add participant information to message flows. To reduce CPU footprint the api is prototyped.
+
+## Breaking:
+- returned api is prototyped
+- second argument passed to extendFn and result of `getExtendContext` is now an instance of ExtendContext. Hence, you are not able to deconstruct `addScript` and/or `addTimers` functions anymore.
+
+## Additions:
+- add participant id and name to `bpmn:MessageFlow`
+- elements are exposed in new api elements property
+- expose participants in elements property
+- add lane information to activity
+>>>>>>> refactor to prototype
 
 # 1.1.0